Multiple Changes
1. Restored deleted files from commit aa173ef as it causes issues when trying to run the server
2. Updated README
3. Integrated Unicode.js and Checker.js files to web app
4. Created a rudimentary GUI layout for basic testing, please change this however you like. The only goal of the commit is to prepare the backend of the web app.

Note: The requirement (as how I understood it) states that UTF32 must be formatted as: xx xx xx where each hex is a nibble,but I forgot that when I was making the backend of the app. Thus prior to output,UTF-32 must be reformatted accordingly. Will try to clarify with Sir so that the formattingcould be done within Unicode.js itself.